Add tools from 1.30.1 release
Add tools from 1.30.1 release as agreed at the 22nd Quattor Workshop,
and integrate with Automake/Autoconf.

The tools are:

   * aqnet
   * aqs
   * ckey
   * diffprof
   * diffuniq (formerly git-diffuniq)
   * getprof
   * prdiff
   * razor
   * rifle (formerly xmlnm)
   * trifle

Each tool is summarised in the README file and has a corresponding man
page.  A COPYRIGHT file describes copying conditions and general terms
of the software license.

This is an Automake/Autoconf project and follows their normal pattern
described in the INSTALL file (./configure; make; make check;
make install; make uninstall).  By default the tools are installed
under /usr/local, but you can pass the configure script a --prefix
option if you want to build it somewhere else.

This project does not produce an installable package.  If one were to
be required, it could be done easily: build to a temporary location
and package up from there.

The trifle tool requires a C compiler and the zlib library and header
files.  If your zlib library and header files are not in a standard
location you can help configure find them using the ZLIB_DIR and
ZLIB_INC environment variables, e.g.

    $ ./configure ZLIB_DIR=<path_to_zlib> ZLIB_INC=<path_to_headers>