getprof now supports downloading multiple profiles
Add -d option to getprof so that an output directory may be supplied
where the downloaded profile can be stored, rather than always sending
it to stdout.

Can now provide a list of hostnames or cluster names to the getprof tool,
and when coupled with the -d option will download multiple profiles to
the output directory at once.

Cluster member profiles can be automatically included in the download
with the -m option, when coupled with -C.